Add files via upload
Cognos extension to make keyboard inputs more useful. Cognos Interactive reports don't let you press "Enter" or "Return" to submit/Finish a prompt page. This extensions allows you to press enter and it executes the Finish request from a custom control that you point to the javascript file found in the extensions folder. 

Secondly, this changes all text on the page and changes it's tabindex to "-1". This makes it easier to Keyboard Tab from one empty field to another assuming you only have text and prompts on the page. This has made keyboard inputs much faster and useful for users since they don't need to use the mouse to submit any part of the prompting.